test: extract TelemetryTestSupport (telemetry/privacy-data mocks) from TestSupport (#554, #434)
Fourth slice of splitting the large TestSupport.swift into domain files,
one cohesive mock group per PR, names stable, green between. Slices
434a-c (Recording, Screenshot, Export) shipped in #551/#558/#559.

Moves the two telemetry mocks — MockOperationalTelemetryRecorder
(OperationalTelemetryRecording) and MockLocalPrivacyDataManager
(LocalPrivacyDataManaging) — byte-identical from TestSupport.swift into
a new TelemetryTestSupport.swift with the minimum imports (Foundation
plus @testable import BugNarrator). Both mocks are self-contained
conformance stubs; MockLocalPrivacyDataManager's MainActor isolation
flows from the @MainActor annotation on the LocalPrivacyDataManaging
protocol itself and is preserved without adding a redundant annotation
on the mock.

Every other mock (MockHotkeyManager, MockURLHandler, MockKeychainService,
actor mocks, etc.) and AppStateHarness stay in TestSupport.swift. Pure
test-support move: no product code, no helper renames, no behavior
change. The full BugNarratorTests suite stays green at 662 tests
(unchanged); docs-sync validation passes.

Pre-build spec reviewed by codex in a fresh session (5 findings folded
in: current 300-330 lines named by protocol not line number, no
annotation changes, minimum imports, allowed-files list, exact 662
baseline).
